:memo: Add documentation comments on `autocmd`
