+++ denops_std/autocmd/common.ts
@@ -1,23 +1,44 @@
 import { Denops } from "https://deno.land/x/denops_core@v3.4.1/mod.ts";
-import { AutocmdEvent } from "./types.ts";
+import {
+  AutocmdEvent,
+  DefineOptions,
+  EmitOptions,
+  ListOptions,
+  RemoveOptions,
+} from "./types.ts";
+import { buildDefineExpr, buildRemoveExpr } from "./utils.ts";
 
-type CommonOptions = {
-  group?: string;
-};
-
-export type DefineOptions = CommonOptions & {
-  once?: boolean;
-  nested?: boolean;
-};
-
-export type RemoveOptions = CommonOptions;
-
-export type ListOptions = CommonOptions;
-
-export type EmitOptions = CommonOptions & {
-  nomodeline?: boolean;
-};
-
+/**
+ * Define an autocmd
+ *
+ * ```typescript
+ * import { Denops } from "../mod.ts";
+ * import * as autocmd from "./mod.ts";
+ *
+ * export async function main(denops: Denops): Promise<void> {
+ *   // Define new autocmd for BufEnter
+ *   await autocmd.define(denops, "BufEnter", "*", "echo 'BufEnter'");
+ *
+ *   // Define new autocmd for BufEnter in 'MyGroup'
+ *   await autocmd.define(denops, "BufEnter", "*", "echo 'BufEnter'", {
+ *     group: "MyGroup",
+ *   });
+ *
+ *   // Define new autocmd for BufEnter with '++once'
+ *   await autocmd.define(denops, "BufEnter", "*", "echo 'BufEnter'", {
+ *     once: true,
+ *   });
+ *
+ *   // Define new autocmd for BufEnter with '++nested'
+ *   await autocmd.define(denops, "BufEnter", "*", "echo 'BufEnter'", {
+ *     nested: true,
+ *   });
+ *
+ *   // Define multiple autocmds
+ *   await autocmd.define(denops, ["BufEnter", "WinEnter"], "*", "echo 'Enter'");
+ * }
+ * ```
+ */
 export async function define(
   denops: Denops,
   event: AutocmdEvent | AutocmdEvent[],
@@ -29,6 +50,30 @@ export async function define(
   await denops.cmd(expr);
 }
 
+/**
+ * Remove an autocmd
+ *
+ * ```typescript
+ * import { Denops } from "../mod.ts";
+ * import * as autocmd from "./mod.ts";
+ *
+ * export async function main(denops: Denops): Promise<void> {
+ *   // Remove BufEnter autocmd
+ *   await autocmd.remove(denops, "BufEnter", "*");
+ *
+ *   // Remove any autocmd in buffer
+ *   await autocmd.remove(denops, "*", "<buffer>");
+ *
+ *   // Remove BufEnter autocmd in 'MyGroup'
+ *   await autocmd.remove(denops, "BufEnter", "*", {
+ *     group: "MyGroup",
+ *   });
+ *
+ *   // Remove multiple autocmds
+ *   await autocmd.remove(denops, ["BufEnter", "WinEnter"], "*");
+ * }
+ * ```
+ */
 export async function remove(
   denops: Denops,
   event?: "*" | AutocmdEvent | AutocmdEvent[],
@@ -39,6 +84,28 @@ export async function remove(
   await denops.cmd(expr);
 }
 
+/**
+ * List defined autocmds
+ *
+ * ```typescript
+ * import { Denops } from "../mod.ts";
+ * import * as autocmd from "./mod.ts";
+ *
+ * export async function main(denops: Denops): Promise<void> {
+ *   // List all autocmd
+ *   console.log(await autocmd.list(denops));
+ *
+ *   // List all BufEnter autocmd
+ *   console.log(await autocmd.list(denops, "BufEnter"));
+ *
+ *   // List all BufEnter autocmd in buffer
+ *   console.log(await autocmd.list(denops, "BufEnter", "<buffer>"));
+ *
+ *   // List multiple autocmds
+ *   console.log(await autocmd.list(denops, ["BufEnter", "WinEnter"]));
+ * }
+ * ```
+ */
 export async function list(
   denops: Denops,
   event?: "*" | AutocmdEvent | AutocmdEvent[],
@@ -67,6 +134,25 @@ export async function list(
   return await denops.call("execute", expr);
 }
 
+/**
+ * Emit an autocmd in a buffer
+ *
+ * ```typescript
+ * import { Denops } from "../mod.ts";
+ * import * as autocmd from "./mod.ts";
+ *
+ * export async function main(denops: Denops): Promise<void> {
+ *   // Emit an autocmd in a current buffer
+ *   await autocmd.emit(denops, "BufEnter");
+ *
+ *   // Emit multiple autocmds in a current buffer
+ *   await autocmd.emit(denops, ["BufEnter", "WinEnter"]);
+ *
+ *   // Emit an autocmd in a buffer 'Hello'
+ *   await autocmd.emit(denops, "BufEnter", "Hello");
+ * }
+ * ```
+ */
 export async function emit(
   denops: Denops,
   event: AutocmdEvent | AutocmdEvent[],
@@ -92,6 +178,25 @@ export async function emit(
   return await denops.cmd(expr);
 }
 
+/**
+ * Emit an autocmd in all buffers
+ *
+ * ```typescript
+ * import { Denops } from "../mod.ts";
+ * import * as autocmd from "./mod.ts";
+ *
+ * export async function main(denops: Denops): Promise<void> {
+ *   // Emit an autocmd in all buffers
+ *   await autocmd.emitAll(denops, "BufEnter");
+ *
+ *   // Emit multiple autocmds in all buffers
+ *   await autocmd.emitAll(denops, ["BufEnter", "WinEnter"]);
+ *
+ *   // Emit an autocmd in all buffers match with 'Hello'
+ *   await autocmd.emitAll(denops, "BufEnter", "Hello");
+ * }
+ * ```
+ */
 export async function emitAll(
   denops: Denops,
   event: AutocmdEvent | AutocmdEvent[],
@@ -115,60 +220,4 @@ export async function emitAll(
   }
   const expr = terms.join(" ");
   return await denops.cmd(expr);
-}
-
-export function buildDefineExpr(
-  event: AutocmdEvent | AutocmdEvent[],
-  pat: string | string[],
-  cmd: string,
-  options: DefineOptions = {},
-): string {
-  const terms = ["au"];
-  if (options.group) {
-    terms.push(options.group);
-  }
-  if (Array.isArray(event)) {
-    terms.push(event.join(","));
-  } else {
-    terms.push(event);
-  }
-  if (Array.isArray(pat)) {
-    terms.push(pat.join(","));
-  } else {
-    terms.push(pat);
-  }
-  if (options.once) {
-    terms.push("++once");
-  }
-  if (options.nested) {
-    terms.push("++nested");
-  }
-  terms.push(cmd);
-  return terms.join(" ");
-}
-
-export function buildRemoveExpr(
-  event?: "*" | AutocmdEvent | AutocmdEvent[],
-  pat?: string | string[],
-  options: RemoveOptions = {},
-): string {
-  const terms = ["au!"];
-  if (options.group) {
-    terms.push(options.group);
-  }
-  if (event) {
-    if (Array.isArray(event)) {
-      terms.push(event.join(","));
-    } else {
-      terms.push(event);
-    }
-    if (pat) {
-      if (Array.isArray(pat)) {
-        terms.push(pat.join(","));
-      } else {
-        terms.push(pat);
-      }
-    }
-  }
-  return terms.join(" ");
 }
