+++ denops_std/autocmd/mod.ts
@@ -1,12 +1,17 @@
-export type { AutocmdEvent } from "./types.ts";
-
-export { define, emit, emitAll, list, remove } from "./common.ts";
+/**
+ * A module to provide helper functions to manage `autocmd`.
+ *
+ * @module
+ */
 export type {
+  AutocmdEvent,
   DefineOptions,
   EmitOptions,
   ListOptions,
   RemoveOptions,
-} from "./common.ts";
+} from "./types.ts";
+
+export { define, emit, emitAll, list, remove } from "./common.ts";
 
 export type {
   GroupDefineOptions,
